refactor: LadderGameCommand 분리
